+++ LICENSE
@@ -359,7 +359,7 @@ Exhibit A - Source Code Form License Notice
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at https://mozilla.org/MPL/2.0/.
   
-  Copyright 2021-2024, Motagamwala Taha Arif Ali
+  Copyright 2021-2025, Motagamwala Taha Arif Ali
 
 If it is not possible or desirable to put the notice in a particular
 file, then You may include the notice in a location (such as a LICENSE
