+++ src/Secweb/CrossOriginEmbedderPolicy/__init__.py
