+++ src/Secweb/CrossOriginResourcePolicy/__init__.py
